+++ integration_tests/tests/integration_tests/test_view_integrated_react/__main__.py
@@ -15,7 +15,7 @@ if __name__ == "__main__":
         "case_failure",
         "case_task_and_element",
         "case_log",
-        "case_filter",
+        "case_search",
         "case_big_structures",
     ]